requirements.txt
kashgari-tf==0.5.3
tensorflow==1.14.0